// ClientApp/src/models/GraphData.ts
export type GraphData = {
    photoUrl: string | undefined;
    id : string;
    givenName : string;
    surname: string;
    userPrincipalName: string;
    jobTitle: string;
    mail: string;
    officeLocation: string;
    mobilePhone: string;
    preferredLanguage: string;
    displayName: string;
    businessPhones:[];
    imageUrl: string;
    username: string;
    name: string;
    tenantId: string;
    homeAccountId: string;
    localAccountId: string;
    environment: string;
    idTokenClaims: {};
}